refactor(schema): Split validation.rs into modular directory structure
Refactors the monolithic validation.rs (~3500 lines) into a modular
directory structure for better maintainability and context management.

## New Structure
```
hive-schema/src/validation/
├── mod.rs        - Common types (ValidationError) and re-exports
├── actuator.rs   - ActuatorSpec validators (19 tests)
├── capability.rs - CapabilityAdvertisement validators (5 tests)
├── command.rs    - HierarchicalCommand validators (7 tests)
├── core.rs       - Basic validators (capability, node, cell) (4 tests)
├── effector.rs   - EffectorSpec validators (17 tests)
├── model.rs      - ModelDeployment validators (21 tests)
├── product.rs    - AI/ML Product validators (10 tests)
├── sensor.rs     - SensorSpec validators (15 tests)
└── track.rs      - Track/TrackUpdate validators (7 tests)
```

## Benefits
- Each module focuses on one message type
- Smaller files fit better in editor/LLM context windows
- Improved code organization and discoverability
- No breaking changes to public API (all re-exported)

## Test Results
All 110 validation tests pass
